[finishes 162825103] Add test case and 'retrieve all meetups' controller logic
