+++ server/app.js
@@ -11,8 +11,8 @@ app.set('json spaces', 2);
 app.use(logger('dev'));
 app.use(helmet());
 
+app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
 
 app.use('/api/v1/', meetupAPI);
